xmlConverter knows to ignore builtin functions and modules
There was a bug caused by the xmlConverter trying to parse builtin
functions. This caused it to blow up as the sourcecode for these
functions wasn't available. Now, I've made a list of built-in/installed
functions and modules that will not be parsed.
The downside to this is that an installed module (say) from PyPI will
likely have sourcecode, but will still not be parsed. Such cases are
not yet handled, but I have to work on my thesis now, so I'll worry
about this case later