Function Calls are Handled
Function calls are now handled. Getting the exit lines from within a
function call scope was tricky, but this has been handled.
Also, it's important to note that the first line of execution of a
called funciton is the first statement in the function, not the line
containing the function def. The CFG is generated as such